Fix rem-to-dip
Started trying to use this mixin then noticed a logic error.

The unit of `$size` will always be(!em || !rem) because if it's one it's not the other. Changed to `and` because I think that's the desired functionality? Not entirely certain what the fallback case is for, it's not clear from the commit history.